:sparkles: #64 - ref: refactored the DestructionListCreate
